+++ src/govuk/components/date-input/date-input.test.js
@@ -1,3 +1,5 @@
+import React from 'react';
+import { render } from '@testing-library/react';
 import examples from './examples';
 import { DateInput } from '.';
 import { diffComponentAgainstReferenceNunjucks } from '../../../../tests/utils/govuk-frontend-diff';
@@ -13,3 +15,16 @@ examples.examples.forEach(example => {
 });
 
 diffComponentAgainstReferenceNunjucks('date-input', DateInput, examples);
+
+describe('date input', () => {
+  it('correctly assigns a ref', () => {
+    const data = { ...examples.examples[0].data };
+    const ref = React.createRef();
+
+    data.items[0].ref = ref;
+
+    const { container } = render(<DateInput {...data} />);
+
+    expect(ref.current).toEqual(container.querySelector('input'));
+  });
+});