+++ src/govuk/components/file-upload/file-upload.test.js
@@ -1,3 +1,5 @@
+import React from 'react';
+import { render } from '@testing-library/react';
 import examples from './examples';
 import { FileUpload } from '.';
 import { diffComponentAgainstReferenceNunjucks } from '../../../../tests/utils/govuk-frontend-diff';
@@ -10,3 +12,14 @@ examples.examples.push({
 });
 
 diffComponentAgainstReferenceNunjucks('file-upload', FileUpload, examples);
+
+describe('file upload', () => {
+  it('correctly assigns a ref', () => {
+    const ref = React.createRef();
+    const { container } = render(
+      <FileUpload {...examples.examples[0].data} ref={ref} />
+    );
+
+    expect(ref.current).toEqual(container.querySelector('input'));
+  });
+});