+++ src/govuk/components/radios/radios.story.js
@@ -3,6 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { Radios as RadiosComponent } from '.';
 import examples from './examples';
 import processExampleData from '../../../../utils/processExampleData';
+import { WithItemRefs } from '../../../../utils/WithRef';
 
 const stories = storiesOf('radios', module);
 
@@ -28,3 +29,7 @@ for (const example of Object.values(
 )) {
   stories.add(example.name, () => <Radios {...example.data} />);
 }
+
+stories.add('with ref', () => (
+  <WithItemRefs Component={Radios} {...examples.examples[0].data} />
+));